+++ sdl2-sonic-drivers/src/hardware/opl/woody/WoodyOPL.hpp
@@ -20,7 +20,7 @@ namespace hardware
 
                 virtual bool init() override;
                 virtual void reset() override;
-                virtual void write(const int port, const int val) noexcept override;
+                virtual void write(const uint32_t port, const uint8_t val) noexcept override;
                 virtual uint8_t read(const int port) noexcept override;
                 virtual void writeReg(const int r, const int v) noexcept override;
